one new enemy, the turret
this seems like a pretty big change, but it's really not. 90% of this push is two packages imported from the asset store. there might be some leftovers from the bird enemy in here, if there is don't use them. the bird is unfinished. the turret enemy is just a modified base enemy who cant move and shoots better.