small changes here and there
fixed a few bugs, like bieng able to pick up keys with spells, enemy 2 not being able to be hit by lightning, and the spawn boxes in level 4 deleting projectiles. made the lightning spell hit harder, increased its speed, and gave it a bigger hitbox so it feels much better. increased the hitbox of the keys. slightly increased player base speed so kevels dont feel so big. added elements to the pause menu that describe the objective and what utility spell you have.